Reorganized tests and add findNodesAtLocation test
